requirements.txt
streamlit==1.45.0
langchain==0.3.25
langchain-community==0.3.23
openai==1.77.0
python-dotenv==1.1.0
youtube-transcript-api==1.0.3
tiktoken==0.9.0 